Upgrades.
The weight of the dev tooling is dragging this action down!

I've upgraded it this time but I think we should remove it next time
it's a pain. We have an audit error due to node-fetch which is only a
dependency due to (unbelievably) a lint rule.

I've updated the eslint rules to match what's in the template repo as
the rules were incompatible with the upgrade. HOWEVER they seem to have
changed their minds on semi-colons. Changing every line is a step too
far for me so I've flipped that from never -> always.

Urgh.